+++ Classes/Service/Items/Parents/ParentsService.php
@@ -103,7 +103,7 @@ class ParentsService extends AbstractService implements ObjectConversionAfterSer
             $path .= $pathSuffix;
 
             if (true === Core::get()->getParentsUtility()->classUsesParentsTrait($property)) {
-                $this->objectsWithParentsPaths[$path] = $path;
+                $this->objectsWithParentsPaths[] = $path;
             }
         }
     }
@@ -152,9 +152,9 @@ class ParentsService extends AbstractService implements ObjectConversionAfterSer
     /**
      * Internal function to fill the parents.
      *
-     * @param mixed $entity
-     * @param array $path
-     * @param array $parents
+     * @param mixed    $entity
+     * @param array    $path
+     * @param object[] $parents
      */
     protected function insertParents($entity, array $path, array $parents)
     {
@@ -166,7 +166,7 @@ class ParentsService extends AbstractService implements ObjectConversionAfterSer
                 && Core::get()->getParentsUtility()->classUsesParentsTrait($propertyValue)
             ) {
                 /** @var ParentsTrait $propertyValue */
-                $propertyValue->setParents($parents);
+                $propertyValue->attachParents($this->filterParents($parents));
             }
         } else {
             if (is_object($propertyValue)) {
@@ -176,5 +176,39 @@ class ParentsService extends AbstractService implements ObjectConversionAfterSer
             array_shift($path);
             $this->insertParents($propertyValue, $path, $parents);
         }
+    }
+
+    /**
+     * This function will filter a given array of objects, by removing
+     * unnecessary parents that use `ParentsTrait` and are followed by another
+     * parent that does the same: there is no need to have the full list of
+     * chained parents
+     *
+     * @param object[] $parents
+     * @return object[]
+     */
+    protected function filterParents(array $parents)
+    {
+        $filteredParents = [];
+        $lastParentWithTrait = null;
+
+        foreach ($parents as $parent) {
+            if (Core::get()->getParentsUtility()->classUsesParentsTrait($parent)) {
+                $lastParentWithTrait = $parent;
+            } else {
+                if (null !== $lastParentWithTrait) {
+                    $filteredParents[] = $lastParentWithTrait;
+                    $lastParentWithTrait = null;
+                }
+
+                $filteredParents[] = $parent;
+            }
+        }
+
+        if (null !== $lastParentWithTrait) {
+            $filteredParents[] = $lastParentWithTrait;
+        }
+
+        return $filteredParents;
     }
 }