+++ Tests/Fixture/Model/DummyInterface.php
@@ -1,0 +1,6 @@
+<?php
+namespace Romm\ConfigurationObject\Tests\Fixture\Model;
+
+interface DummyInterface
+{
+}